add a .gitignore file
Object, archive, executable files or any form of tarballs don't need
to be tracked.

.dat, .txt, .out files are either input or output ASCII files that
some routines need, but they don't need to be tracked, because (the
important ones) are tracked/backed up elsewhere

output/ is going to be the directory into which output files should
be written from now on. This still needs to be updated in those
routines where the output path is hard coded in.