new and improved makefile!
the program to be compiled is now specified by passing it to 'PROGS'
when make-ing, instead of in the makefile. This a) makes the makefile
slightly more universal, and b) prevents miscellaneous edits to it to
be tracked in git. Thus, no more uncommenting the appropriate line to
make (despite what most code headers still state)!

Again, when compiling, make sure $TREEHOME is set correctly, and that
the tree libraries have been built once for serial use.