fix counti-inconsistency, add support for both sdf "file formats"
When counting the isotopes for counti, I overcounted by one, and so various
loop conditions needed to be fixed also.

In the old sdf file format, the Z and N information of the tracked isotopes
was in the SPHbody, and N was designated by m1, ... . In the new file format,
the Z and N information was moved to the header (thus ~halving the file size)
and N was designated by 'n1', ..., (and at the same time the number of
tracked isotopes was changed from 22 to 20). Since the read-in of the sdf file
is semi hard-coded in, variables related to that had to be changed by hand
for the different file formats. This is a first attempt at automating this.
Has not been tested yet for the old file format.