bugfix, correct counting of isotopes, instride vs outstride
- fixed a bug that caused the yields-array to contain garbage. Masses
are added to 'yields' cumulatively, thus it needs to be initialized (so
it does not contain junk from memory)
- this is carried over from the last commit, where in one instance the
isotopes were overcounted by one. Some loops dependent on this number
had to be fixed (again)
- SDFseekreadvecsarr technically needs the byte-offset between successive
quantities to be read in; this is 'instride'. If only one line is read in
at a time, this value really has no significance. But this value can
be different from 'outstrides', which just counts the number of bytes in
a line to be written to output. There was (still is) some confusion in the
code between these two.