add option to get (mass) density in cgs of the selected isotopes
to be selected at runtime by setting the log-flag to 3. Mass density
may be a more useful quantity for people not regularly working with
stellar evolution codes of all sorts (it seems). Positions are left in
code units. Other log-flag options were updated (or added, possibly),
and a short description added to the header.