addvel: add option to impose asymmetry in vel and u
this might be necessary in some of the very early explosion models I
get. In some models, most/a majority of the energy is still in internal
energy (u), thus imposing an asymmetry only in velocity (= kinetic
energy) won't have much of an effect. This option can be toggled by
setting the asym_u option to 1 at runtime.