getting mergeSDFs to work again
I think a long time ago I started "improving" it, then decided to
have both a simple and a 'smart' routine for merging SDFs. The
'smart' routine (mergeSDFs2) was then based off the simple one,
which was broken at the time, and never fixed. The 'smart' routine
is far from being smart, or working ....