rm makeabuntable.c, update dmMdvplots
not sure what makeabuntable was; probably me trying to change the name
of a routine, then forgetting about it and continue updating/using the
old file.

increasing number of isotopes in dmMdvplots to 7, which is probably the
max that can reasonably be plotted anyways. also removing some hard-wired
numbers that arose from the number of isotopes being an assumed constant
value.