CSdisk: creating a tool to add a disk of material to an ISM file
the code prompts for the radius and height of the disk (and for a value
for a maximum extend that currently has no effect). it cuts out a disk
from the first file and inserts a disk of material from the second file
into that. particles are renumbered, but not reordered.